+++ Sources/MapboxCoreNavigation/CoreNavigationNavigator.swift
@@ -21,16 +21,21 @@ class Navigator {
      */
     static var tilesURL: URL? = nil
     
-    func enableHistoryRecorder() throws {
-        try historyRecorder.enable(forEnabled: true)
-    }
+    /**
+     Path to the file where history could be stored when `Navigator.dumpHistory(_:)` is called.
+     */
+    static var historyFilePath: URL? = nil
     
-    func disableHistoryRecorder() throws {
-        try historyRecorder.enable(forEnabled: false)
-    }
-    
-    func history() throws -> Data {
-        return try historyRecorder.getHistory()
+    /**
+     Store history to the file stored in `Navigator.historyFilePath` and asynchronously run a callback
+     when dumping finishes.
+     
+     - parameter completion: A block object to be executed when history dumping ends.
+     */
+    func dumpHistory(_ completion: @escaping (String?) -> Void) {
+        historyRecorder.dumpHistory { (path) in
+            completion(path)
+        }
     }
     
     var historyRecorder: HistoryRecorderHandle!
@@ -42,7 +47,7 @@ class Navigator {
     var roadGraph: RoadGraph!
     
     lazy var roadObjectsStore: RoadObjectsStore = {
-        return RoadObjectsStore(try! navigator.roadObjectStore())
+        return RoadObjectsStore(navigator.roadObjectStore())
     }()
     
     /**
@@ -86,29 +91,41 @@ class Navigator {
                                       threadsCount: nil,
                                       endpointConfig: endpointConfig)
         
-        let configFactory = try! ConfigFactory.build(for: settingsProfile,
-                                                     config: NavigatorConfig(),
-                                                     customConfig: "")
+        let historyAutorecordingConfig = [
+            "features": [
+                "historyAutorecording": true
+            ]
+        ]
         
-        historyRecorder = try! HistoryRecorderHandle.build(forHistoryFile: "", config: configFactory)
+        var customConfig = ""
+        if let jsonDataConfig = try? JSONSerialization.data(withJSONObject: historyAutorecordingConfig, options: []),
+           let encodedConfig = String(data: jsonDataConfig, encoding: .utf8) {
+            customConfig = encodedConfig
+        }
         
-        let runloopExecutor = try! RunLoopExecutorFactory.build()
-        cacheHandle = try! CacheFactory.build(for: tilesConfig,
-                                              config: configFactory,
-                                              runLoop: runloopExecutor,
-                                              historyRecorder: historyRecorder)
+        let configFactory = ConfigFactory.build(for: settingsProfile,
+                                                config: NavigatorConfig(),
+                                                customConfig: customConfig)
         
-        roadGraph = RoadGraph(try! MapboxNavigationNative.GraphAccessor(cache: cacheHandle))
+        historyRecorder = HistoryRecorderHandle.build(forHistoryFile: Navigator.historyFilePath?.absoluteString ?? "", config: configFactory)
         
-        navigator = try! MapboxNavigationNative.Navigator(config: configFactory,
-                                                          runLoopExecutor: runloopExecutor,
-                                                          cache: cacheHandle,
-                                                          historyRecorder: historyRecorder)
-        try! navigator.setElectronicHorizonObserverFor(self)
+        let runloopExecutor = RunLoopExecutorFactory.build()
+        cacheHandle = CacheFactory.build(for: tilesConfig,
+                                         config: configFactory,
+                                         runLoop: runloopExecutor,
+                                         historyRecorder: historyRecorder)
+        
+        roadGraph = RoadGraph(MapboxNavigationNative.GraphAccessor(cache: cacheHandle))
+        
+        navigator = MapboxNavigationNative.Navigator(config: configFactory,
+                                                     runLoopExecutor: runloopExecutor,
+                                                     cache: cacheHandle,
+                                                     historyRecorder: historyRecorder)
+        navigator.setElectronicHorizonObserverFor(self)
     }
     
     deinit {
-        try! navigator.setElectronicHorizonObserverFor(nil)
+        navigator.setElectronicHorizonObserverFor(nil)
     }
     
     var electronicHorizonOptions: ElectronicHorizonOptions? {
@@ -119,19 +136,17 @@ class Navigator {
             } else {
                 nativeOptions = nil
             }
-            try! navigator.setElectronicHorizonOptionsFor(nativeOptions)
+            navigator.setElectronicHorizonOptionsFor(nativeOptions)
         }
     }
-    
-    var peer: MBXPeerWrapper?
 }
 
 extension Navigator: ElectronicHorizonObserver {
     public func onPositionUpdated(for position: ElectronicHorizonPosition, distances: [String : MapboxNavigationNative.RoadObjectDistanceInfo]) {
         let userInfo: [ElectronicHorizon.NotificationUserInfoKey: Any] = [
-            .positionKey: RoadGraph.Position(try! position.position()),
-            .treeKey: ElectronicHorizon(try! position.tree()),
-            .updatesMostProbablePathKey: try! position.type() == .UPDATE,
+            .positionKey: RoadGraph.Position(position.position()),
+            .treeKey: ElectronicHorizon(position.tree()),
+            .updatesMostProbablePathKey: position.type() == .UPDATE,
             .distancesByRoadObjectKey: distances.mapValues(RoadObjectDistanceInfo.init),
         ]
         NotificationCenter.default.post(name: .electronicHorizonDidUpdatePosition, object: nil, userInfo: userInfo)