+++ Sources/MapboxCoreNavigation/EHorizon/RoadObjectsStore.swift
@@ -18,9 +18,9 @@ public final class RoadObjectsStore {
     public weak var delegate: RoadObjectsStoreDelegate? {
         didSet {
             if delegate != nil {
-                try! native.setObserverForOptions(self)
+                native.setObserverForOptions(self)
             } else {
-                try! native.setObserverForOptions(nil)
+                native.setObserverForOptions(nil)
             }
         }
     }
@@ -31,7 +31,7 @@ public final class RoadObjectsStore {
      - parameter edgeIdentifier: The identifier of the edge to query.
      */
     public func roadObjectEdgeLocations(edgeIdentifier: ElectronicHorizon.Edge.Identifier) -> [RoadObjectIdentifier : RoadObjectEdgeLocation] {
-        let objects = try! native.getForEdgeId(UInt64(edgeIdentifier))
+        let objects = native.getForEdgeId(UInt64(edgeIdentifier))
         return objects.mapValues(RoadObjectEdgeLocation.init)
     }
 
@@ -42,7 +42,7 @@ public final class RoadObjectsStore {
      - parameter roadObjectIdentifier: The identifier of the road object to query.
      */
     public func roadObjectMetadata(identifier roadObjectIdentifier: RoadObjectIdentifier) -> RoadObjectMetadata? {
-        if let metadata = try! native.getRoadObjectMetadata(forRoadObjectId: roadObjectIdentifier) {
+        if let metadata = native.getRoadObjectMetadata(forRoadObjectId: roadObjectIdentifier) {
             return RoadObjectMetadata(metadata)
         }
         return nil
@@ -55,7 +55,7 @@ public final class RoadObjectsStore {
      - parameter roadObjectIdentifier: The identifier of the road object to query.
      */
     public func roadObjectLocation(identifier roadObjectIdentifier: RoadObjectIdentifier) -> RoadObjectLocation? {
-        if let location = try! native.getRoadObjectLocation(forRoadObjectId: roadObjectIdentifier) {
+        if let location = native.getRoadObjectLocation(forRoadObjectId: roadObjectIdentifier) {
             return RoadObjectLocation(location)
         }
         return nil
@@ -66,17 +66,15 @@ public final class RoadObjectsStore {
      - parameter edgeIds list of edge ids
      */
     public func roadObjectIdentifiers(edgeIdentifiers: [ElectronicHorizon.Edge.Identifier]) -> [RoadObjectIdentifier] {
-        return try! native.getRoadObjectIdsByEdgeIds(forEdgeIds: edgeIdentifiers.map(NSNumber.init))
+        return native.getRoadObjectIdsByEdgeIds(forEdgeIds: edgeIdentifiers.map(NSNumber.init))
     }
-
-    public var peer: MBXPeerWrapper?
 
     init(_ native: MapboxNavigationNative.RoadObjectsStore) {
         self.native = native
     }
 
     deinit {
-        try! native.setObserverForOptions(nil)
+        native.setObserverForOptions(nil)
     }
 
     private let native: MapboxNavigationNative.RoadObjectsStore