+++ Sources/MapboxCoreNavigation/RouteController.swift
@@ -222,16 +222,16 @@ open class RouteController: NSObject {
         let activeGuidanceOptions = ActiveGuidanceOptions(mode: mode(progress.routeOptions.profileIdentifier),
                                                           geometryEncoding: geometryEncoding(progress.routeOptions.shapeFormat),
                                                           waypoints: waypoints)
-        try! navigator.setRouteForRouteResponse(routeJSONString,
-                                                route: 0,
-                                                leg: UInt32(routeProgress.legIndex),
-                                                options: activeGuidanceOptions)
+        navigator.setRouteForRouteResponse(routeJSONString,
+                                           route: 0,
+                                           leg: UInt32(routeProgress.legIndex),
+                                           options: activeGuidanceOptions)
     }
     
     /// updateRouteLeg is used to notify nav-native of the developer changing the active route-leg.
     private func updateRouteLeg(to value: Int) {
         let legIndex = UInt32(value)
-        if try! navigator.changeRouteLeg(forRoute: 0, leg: legIndex), let timestamp = location?.timestamp {
+        if navigator.changeRouteLeg(forRoute: 0, leg: legIndex), let timestamp = location?.timestamp {
             updateIndexes(status: navigator.status(at: timestamp), progress: routeProgress)
         }
     }
@@ -245,7 +245,7 @@ open class RouteController: NSObject {
         
         rawLocation = location
         
-        locations.forEach { try! navigator.updateLocation(for: FixLocation($0)) }
+        locations.forEach { navigator.updateLocation(for: FixLocation($0)) }
 
         let status = navigator.status(at: location.timestamp)
         
@@ -384,16 +384,23 @@ open class RouteController: NSObject {
         updateRouteLeg(to: routeProgress.legIndex + 1)
     }
     
-    public func enableLocationRecording() {
-        try! Navigator.shared.enableHistoryRecorder()
+    /**
+     Path to the file where history could be stored when `RouteController.dumpHistory(_:)` is called.
+     */
+    public static var historyFilePath: URL? = nil {
+        didSet {
+            Navigator.historyFilePath = historyFilePath
+        }
     }
     
-    public func disableLocationRecording() {
-        try! Navigator.shared.disableHistoryRecorder()
-    }
-    
-    public func locationHistory() throws -> Data {
-        return try Navigator.shared.history()
+    /**
+     Store history to the file stored in `RouteController.historyFilePath` and asynchronously run a callback
+     when dumping finishes.
+     
+     - parameter completion: A block object to be executed when history dumping ends.
+     */
+    public static func dumpHistory(_ completion: @escaping (String?) -> Void) {
+        Navigator.shared.dumpHistory(completion)
     }
     
     /**