+++ Sources/MapboxNavigation/NavigationMapView.swift
@@ -240,18 +240,23 @@ open class NavigationMapView: UIView {
     }
     
     func setupMapView(_ frame: CGRect, navigationCameraType: NavigationCameraType = .mobile) {
-        guard let accessToken = AccountManager.shared.accessToken else {
+        guard let accessToken = CredentialsManager.default.accessToken else {
             fatalError("Access token was not set.")
         }
         
-        let options = ResourceOptions(accessToken: accessToken,
-                                      tileStorePath: Bundle.mapboxNavigation.suggestedTileURL?.path,
-                                      loadTilePacksFromNetwork: false)
+        var tileStore = TileStore.getInstance()
+        if let suggestedTilePath = Bundle.mapboxNavigation.suggestedTileURL?.path {
+            tileStore = TileStore.getInstanceForPath(suggestedTilePath)
+        }
+        let resourceOptions = ResourceOptions(accessToken: accessToken,
+                                              tileStore: tileStore,
+                                              tileStoreEnabled: true,
+                                              loadTilePacksFromNetwork: false)
         
-        mapView = MapView(with: frame, resourceOptions: options)
+        mapView = MapView(frame: frame, mapInitOptions: MapInitOptions(resourceOptions: resourceOptions))
         mapView.translatesAutoresizingMaskIntoConstraints = false
         mapView.update {
-            $0.ornaments.showsScale = false
+            $0.ornaments.scaleBarVisibility = .hidden
         }
         
         mapView.on(.renderFrameFinished) { [weak self] _ in
@@ -282,11 +287,8 @@ open class NavigationMapView: UIView {
      - parameter options: options, controlling caching parameters like area radius and concurrent downloading threads.
      */
     public func enablePredictiveCaching(options predictiveCacheOptions: PredictiveCacheOptions) {
-        let mapTileSource = try? TileStoreManager.getTileStore(for: mapView.__map.getResourceOptions())
-        var mapOptions: PredictiveCacheManager.MapOptions?
-        if let tileStore = mapTileSource?.value as? TileStore {
-            mapOptions = PredictiveCacheManager.MapOptions(tileStore, mapView.styleSourceDatasets(["raster", "vector"]))
-        }
+        let mapOptions = PredictiveCacheManager.MapOptions(TileStore.getInstance(),
+                                                           mapView.styleSourceDatasets(["raster", "vector"]))
         
         predictiveCacheManager = PredictiveCacheManager(predictiveCacheOptions: predictiveCacheOptions,
                                                         mapOptions: mapOptions)
@@ -352,8 +354,8 @@ open class NavigationMapView: UIView {
         // While animating to overview mode, don't animate the puck.
         let duration: TimeInterval = animated && navigationCamera.state != .transitionToOverview ? 1 : 0
         UIView.animate(withDuration: duration, delay: 0, options: [.curveLinear]) { [weak self] in
-            guard let point = self?.mapView.screenCoordinate(for: location.coordinate).point else { return }
-            self?.userCourseView.center = point
+            guard let screenCoordinate = self?.mapView.screenCoordinate(for: location.coordinate) else { return }
+            self?.userCourseView.center = CGPoint(x: screenCoordinate.x, y: screenCoordinate.y)
         }
         
         userCourseView.update(location: location,
@@ -390,9 +392,9 @@ open class NavigationMapView: UIView {
     func fitCamera(to route: Route, animated: Bool = false) {
         guard let routeShape = route.shape, !routeShape.coordinates.isEmpty else { return }
         let edgeInsets = safeArea + UIEdgeInsets(top: 10.0, left: 20.0, bottom: 10.0, right: 20.0)
-        if let cameraOptions = mapView?.cameraManager.camera(fitting: .lineString(routeShape),
-                                                             edgePadding: edgeInsets) {
-            mapView?.cameraManager.setCamera(to: cameraOptions, animated: animated)
+        if let cameraOptions = mapView?.camera.camera(fitting: .lineString(routeShape),
+                                                      edgePadding: edgeInsets) {
+            mapView?.camera.setCamera(to: cameraOptions, animated: animated)
         }
     }
     
@@ -433,7 +435,7 @@ open class NavigationMapView: UIView {
                                                 coordinate.latitude,
                                                 mapView.bounds.size))
         
-        mapView.cameraManager.setCamera(to: CameraOptions(center: coordinate, zoom: zoom))
+        mapView.cameraOptions = CameraOptions(center: coordinate, zoom: zoom)
         updateUserCourseView(CLLocation(latitude: coordinate.latitude, longitude: coordinate.longitude))
     }
     
@@ -470,10 +472,9 @@ open class NavigationMapView: UIView {
                 IdentifierString.buildingExtrusionLayer
             ]
             
-            guard let layers = try? mapView.__map.getStyleLayers().reversed() else { return nil }
-            for layer in layers {
+            for layer in mapView.mapboxMap.__map.getStyleLayers().reversed() {
                 if !(layer.type == "symbol") && !identifiers.contains(layer.id) {
-                    let sourceLayer = try? mapView.__map.getStyleLayerProperty(forLayerId: layer.id, property: "source-layer").value as? String
+                    let sourceLayer = mapView.mapboxMap.__map.getStyleLayerProperty(forLayerId: layer.id, property: "source-layer").value as? String
                     
                     if let sourceLayer = sourceLayer,
                        sourceLayer.isEmpty {
@@ -611,11 +612,11 @@ open class NavigationMapView: UIView {
         }
 
         if let lastLeg = route.legs.last, let destinationCoordinate = lastLeg.destination?.coordinate {
-            mapView.annotationManager.removeAnnotations(annotationsToRemove())
+            mapView.annotations.removeAnnotations(annotationsToRemove())
             
             var destinationAnnotation = PointAnnotation(coordinate: destinationCoordinate)
             destinationAnnotation.title = "navigation_annotation"
-            mapView.annotationManager.addAnnotation(destinationAnnotation)
+            mapView.annotations.addAnnotation(destinationAnnotation)
         }
     }
     
@@ -713,7 +714,7 @@ open class NavigationMapView: UIView {
     }
     
     public func removeWaypoints() {
-        mapView.annotationManager.removeAnnotations(annotationsToRemove())
+        mapView.annotations.removeAnnotations(annotationsToRemove())
         
         let layerSet: Set = [IdentifierString.waypointCircle,
                              IdentifierString.waypointSymbol]
@@ -723,7 +724,7 @@ open class NavigationMapView: UIView {
     
     func annotationsToRemove() -> [Annotation] {
         // TODO: Improve annotations filtering functionality.
-        return mapView.annotationManager.annotations.values.filter({ $0.title == "navigation_annotation" })
+        return mapView.annotations.annotations.values.filter({ $0.title == "navigation_annotation" })
     }
     
     /**
@@ -793,8 +794,8 @@ open class NavigationMapView: UIView {
             if let _ = try? mapView.style.getSource(identifier: IdentifierString.arrowSymbolSource, type: GeoJSONSource.self).get() {
                 let geoJSON = Feature.init(geometry: Geometry.point(point))
                 let _ = mapView.style.updateGeoJSON(for: IdentifierString.arrowSymbolSource, with: geoJSON)
-                let _ = try? mapView.__map.setStyleLayerPropertyForLayerId(IdentifierString.arrowSymbol, property: "icon-rotate", value: shaftDirection)
-                let _ = try? mapView.__map.setStyleLayerPropertyForLayerId(IdentifierString.arrowCasingSymbol, property: "icon-rotate", value: shaftDirection)
+                mapView.mapboxMap.__map.setStyleLayerPropertyForLayerId(IdentifierString.arrowSymbol, property: "icon-rotate", value: shaftDirection)
+                mapView.mapboxMap.__map.setStyleLayerPropertyForLayerId(IdentifierString.arrowCasingSymbol, property: "icon-rotate", value: shaftDirection)
             } else {
                 var arrowSymbolLayer = SymbolLayer(id: IdentifierString.arrowSymbol)
                 arrowSymbolLayer.minZoom = Double(minimumZoomLevel)